restructure backupstore package
